Commit Initial HTML UI and Handlebars for goals
I added html forms in index.html to test the ability to create and
update goals.  Added a table to display our user's goal data.  Created
an initial handlebar template to generate user goals and generate delete
and modify buttons for each entry. Will need to add to JS to render and
test.