Fixed Update Goal Form Functionality
Handlebars Changes
Changes made to handlebars template to add ids to individual table data
for easy access.

HTML Changes
Messaging appears if user tries to update a record without changes.

JS Changes
When "modify goal" form loads, field values now reflect the values
rendered in the table and in the backend database.

Fixed form behavior if user decides to change the record to modify while
already in the process of modifying the form.